atualizacao geral

+++ controllers/UserController.js
@@ -13,12 +13,11 @@ const retrieveUsers = (req, res) => {
 const createUsers = (req, res) => {
   // Validação pode ser adicionada aqui
   sql.query(
-    "INSERT INTO Users (UserName, Email, Pass, Rol) VALUES (?,?,?,?)",
+    "INSERT INTO Users (UserName, Email, Pass) VALUES (?,?,?)",
     [
       req.body.UserName,
       req.body.Email,
-      req.body.Pass,
-      req.body.Rol
+      req.body.Pass
     ],
     function (err, result) {
       if (err) {
@@ -53,35 +52,39 @@ const retrieveUser = (req, res) => {
   );
 };
 
-const deleteUsers = (req, res) => {
-  console.log("UserID to delete:", req.params.ID);
+const deleteUsers = async (req, res) => {
+  const userId = req.params.ID; // Obtém o ID do usuário a partir dos parâmetros da requisição
 
-  sql.query(
-    "DELETE FROM Users WHERE ID = ?",
-    [req.params.ID],
-    function (err, result) {
-      if (err) {
-        console.error("Error deleting user:", err);
-        return res.status(500).send("Error deleting user");
-      }
-      if (result.affectedRows === 0) {
-        res.status(404).send("User not found");
-      } else {
-        res.send("User " + req.params.ID + " successfully deleted");
-      }
-    }
-  );
+  try {
+      // Iniciar uma transação
+      await sql.beginTransaction();
+
+      // Excluir registros relacionados na tabela Cart
+      await sql.query("DELETE FROM Cart WHERE UserID = ?", [userId]);
+
+      // Excluir o usuário
+      await sql.query("DELETE FROM Users WHERE ID = ?", [userId]);
+
+      // Confirmar a transação
+      await sql.commit();
+      
+      res.status(200).send("Usuário excluído com sucesso.");
+  } catch (error) {
+      // Reverter a transação em caso de erro
+      await sql.rollback();
+      console.error("Erro ao excluir o usuário:", error);
+      res.status(500).send("Erro ao excluir o usuário.");
+  }
 };
 
 const updateUsers = (req, res) => {
   // Validação pode ser adicionada aqui
   sql.query(
-    "UPDATE Users SET UserName = ?, Email = ?, Pass = ?, Rol = ? WHERE ID = ?",
+    "UPDATE Users SET UserName = ?, Email = ?, Pass = ? WHERE ID = ?",
     [
       req.body.UserName,
       req.body.Email,
       req.body.Pass,
-      req.body.Rol,
       req.params.ID
     ],
     function (err, result) {
